### 2018-11-27 c50b7e8
Update paper.md
