### 2018-11-27 #1099 @arfon
@melund - I think this OR should fix your paper: https://github.com/AnyBody-Research-Group/AnyPyTools/pull/55
