### 2018-11-27 #1099 @melund
Thanks, @arfon. I merged AnyBody-Research-Group/AnyPyTools#55
