### 2018-11-27 #1099 @melund
@trallard Most of the review can probably be handled by someone who knows Python/GitHub/OpenSource.

But having reviewers with knowledge on multibody musculoskeletal modeling and the [AnyBody Modeling System](https://www.anybodytech.com/software/ams/) is probably also required.

So here are a few suggestions on reviewers who have used the AnyBody Modeling System, and have some Python and GitHub/Git experience.

1. [Antonio Gonzales Marin](https://www.linkedin.com/in/antonio-gonzales-marin-817245115/) (@tonygon). Research Scientist at the German Aerospace Center (DLR)

2. [Marco Antonio Marra](https://www.linkedin.com/in/marcoantoniomarra/) (@marcopus) Assistant Professor at University of Twente
