### 2018-11-28 #1099 @trallard
👋 Hi @melund thanks for the submission. Since you suggested me as an editor I will be performing the editorial duties for this submission. If you have any questions over the review process please feel free to ping me.

Let's start by pinging some suitable reviewers with Python and biomechanics expertise. We can start the review once two or more of them accept the request to review.

---
Dear reviewers,

@tonygon as suggested: Antonio, are you able to and interested in taking on this review?
@marcopus as suggested: Marco, are you able to and interested in taking on this review?
@moorepants: Jason, are you able to and interested in taking on this review?
@nirum: Niru, are you able to and interested in taking on this review?
@acolum: Alyssa, are you able to and interested in taking on this review?

**Title**: AnyPyTools: A Python package for reproducible research with the AnyBody Modeling System.
**Summary**: AnyPyTools is a toolkit for working with the AnyBody Modeling System (AMS) from Python. It enables reproducible research with the AnyBody Modeling System and bridges the gap to the whole ecosystem of open source scientific Python.
**Topic**: reproducible science, musculoskeletal modelling
**Article proof**: [https://github.com/openjournals/joss-papers/blob/joss.01099/joss.01099/10.21105.joss.01099.pdf](https://github.com/openjournals/joss-papers/blob/joss.01099/joss.01099/10.21105.joss.01099.pdf)
**Submitting author**: @melund
**Repository**: [https://github.com/AnyBody-Research-Group/AnyPyTools](https://github.com/AnyBody-Research-Group/AnyPyTools)
