### 2018-11-28 #1099 @marcopus
Hi Tania, I am willing to review this. Thanks, Marco

Il giorno mer 28 nov 2018 alle ore 15:34 Tania Allard <
notifications@github.com> ha scritto:

> 👋 Hi @melund <https://github.com/melund> thanks for the submission.
> Since you suggested me as an editor I will be performing the editorial
> duties for this submission. If you have any questions over the review
> process please feel free to ping me.
>
> Let's start by pinging some suitable reviewers with Python and
> biomechanics expertise. We can start the review once two or more of them
> accept the request to review.
> ------------------------------
>
> Dear reviewers,
>
> @tonygon <https://github.com/tonygon> as suggested: Antonio, are you able
> to and interested in taking on this review?
> @marcopus <https://github.com/marcopus> as suggested: Marco, are you able
> to and interested in taking on this review?
> @moorepants <https://github.com/moorepants>: Jason, are you able to and
> interested in taking on this review?
> @nirum <https://github.com/nirum>: Niru, are you able to and interested
> in taking on this review?
> @acolum <https://github.com/acolum>: Alyssa, are you able to and
> interested in taking on this review?
>
> *Title*: AnyPyTools: A Python package for reproducible research with the
> AnyBody Modeling System.
> *Summary*: AnyPyTools is a toolkit for working with the AnyBody Modeling
> System (AMS) from Python. It enables reproducible research with the AnyBody
> Modeling System and bridges the gap to the whole ecosystem of open source
> scientific Python.
> *Topic*: reproducible science, musculoskeletal modelling
> *Article proof*:
> https://github.com/openjournals/joss-papers/blob/joss.01099/joss.01099/10.21105.joss.01099.pdf
> *Submitting author*: @melund <https://github.com/melund>
> *Repository*: https://github.com/AnyBody-Research-Group/AnyPyTools
>
