### 2018-11-29 #1108 @whedon
[REVIEW]: AnyPyTools: A Python package for reproducible research with the AnyBody Modeling System.
**Submitting author:** @melund (<a href="http://orcid.org/0000-0001-9920-4051">Morten Enemark Lund</a>)
**Repository:** <a href="https://github.com/AnyBody-Research-Group/AnyPyTools" target ="_blank">https://github.com/AnyBody-Research-Group/AnyPyTools</a>
**Version:** 1.0.0
