### 2018-11-29 #1108 @trallard
Hi reviewers and @melund! This is the issue we will be using to conduct the review.
@tonygon, @marcopus, @acolum each of you have a checklist to guide you through the review process.

If any of you encounters any issues / need any guidance during the review process feel free to ping me here 👩🏻‍💻
