### 2018-12-09 #1108 @acolum
Just completed my review!
The paper, documentation, and code were all excellent, and I was glad to check off all of the requirements above.
