### 2018-12-11 #1108 @melund
@trallard. I just browsed the [JOSS conflict of interest policy](https://github.com/openjournals/joss/blob/master/COI.md) and I think one of the reviewers I suggested could cause a problem. I knew one of my supervisors (Michael Skipper Andersen @msna) had worked with @marcopus in the past but didn't realize that the also recently (2017) co-authored a paper.

I am sorry about that. But I think that disqualifies @marcopus as a reviewer.
