### 2018-12-18 #1108 @tonygon
> hey @melund thanks for double checking. I am going to then remove @marcopus (apologies). Is @tonygon still a suitable reviewer?

@trallard My apologies for the delay. As I mentioned to Morten before agreeing, I'm able to do the review, I work with AnyPython tools and AnyBody, I used both regularly for a project.
But, I'm in the transition between jobs, so time is a bit limited ;)
Since we have holidays by the end of the week I will finish the review then.
I already know/did most of all AnyPython tools part, I just need to seat down and finish the formal procedure from the JOSS side.

I hope there is still time.
Cheers!
