### 2018-12-20 87a37f0
Fixed problem with pytest plugin when pytest-xidst is not installed
