### 2018-12-30 #1108 @tonygon
@trallard
Just finished checking the JOSS requirements)
Thanks for the patience!
And continue the great work @melund!!

Cheers!
