### 2019-01-02 #1108 @trallard
@tonygon thanks for your review and hope you had a nice holiday season

---
@melund now that the reviews are completed and the paper has been recommended for acceptance, we need you to complete the following actions:

- [ ]  make a new release of the package (unless the latest release is up to date with the recent modifications)

- [ ]  upload the revised software to your DOI-granting data/software repository, and post the DOI here

- [ ] double check the submission paper to ensure it is up-to-date

Once these tasks are completed I can finalise the editorial tasks
