### 2019-01-02 e9d35cf
Update reference to Zenodo with DOI for all versions
