+++ docs/paper/paper.bib
@@ -3,9 +3,9 @@
   title  = "{AnyPyTools}",
   author = "Lund, Morten Enemark and Andersen, Michael Skipper and Rasmussen,
             John",
-  month  =  oct,
+  month  =  dec,
   year   =  2018,
-  doi    = "10.5281/zenodo.1474200"
+  doi    = "10.5281/zenodo.1251087"
 }
 
 @ARTICLE{De_Pieri2018-py,
