### 2019-01-02 #1108 @melund
Thanks a lot @trallard. I have finished the remaining tasks.

> - [x] make a new release of the package (unless the latest release is up to date with the recent modifications)
>
> - [x] upload the revised software to your DOI-granting data/software repository, and post the DOI here
>
> - [x] double check the submission paper to ensure it is up-to-date

The library is achived with Zenodo and the DOI for version 1.0.1 is:

**10.5281/zenodo.2450121**
