### 2019-01-07 #1108 @trallard
@melund all looks good to me so I will proceed with the acceptance recommendation

@arfon: this submission is accepted and ready to be published 🎉👾

@tonygon @acolum  thank you very much for your time and valuable contribution to JOSS as reviewers for this submission 🙌🏻
